--- docs/.dockerignore
@@ -1,4 +1,0 @@
-.react-router
-build
-node_modules
-README.md
