--- docs/app/routes/about_page.tsx
@@ -1,13 +1,0 @@
-import type { Route } from "../+types/root";
-import { About } from "../about/about";
-
-export function meta({}: Route.MetaArgs) {
-  return [
-    { title: "New React Router App" },
-    { name: "description", content: "Welcome to React Router!" },
-  ];
-}
-
-export default function About_page() {
-  return <About />;
-}
