+++ docs/public/main.css
@@ -1,3 +1,4 @@
+@import "tailwindcss";
 /**
  * Reset some basic elements
  */
