+++ docs/src/routes/about_page.jsx
@@ -1,0 +1,5 @@
+import About from "@/pages/About";
+
+export default function AboutPage(){
+    return <About />
+}
