+++ docs/src/routes/contact_page.jsx
@@ -1,0 +1,5 @@
+import Contact from "@/pages/Contact";
+
+export default function ContactPage(){
+    return <Contact />
+}
