+++ docs/src/routes/home_page.jsx
@@ -1,0 +1,5 @@
+import Home from "@/pages/Home";
+
+export default function HomePage(){
+    return <Home />
+}
