+++ docs/src/routes/post_routes/welcome-to-the-blog_page.jsx
@@ -1,0 +1,5 @@
+import { WelcomeToTheBlog } from "@/pages/posts/Welcome-to-the-blog";
+
+export default function WelcomeToTheBlogPage() {
+    return <WelcomeToTheBlog />;
+}
